refactor duplicated GetDialogCtx()
