+++ src/route_gui.cpp
@@ -12,6 +12,7 @@
 #include "georef.h"
 #include "gui_lib.h"
 #include "multiplexer.h"
+#include "n0183_ctx_factory.h"
 #include "navutil.h"
 #include "own_ship.h"
 #include "routeman.h"
@@ -578,25 +579,6 @@ void RouteGui::CalculateDCRect(wxDC &dc_route, ChartCanvas *canvas,
   }
 
   *prect = update_rect;
-}
-
-static  N0183DlgCtx GetDialogCtx(SendToGpsDlg* dialog) {
-  N0183DlgCtx dlg_ctx;
-  dlg_ctx.set_value = [dialog](int v) {
-      if (!dialog || !dialog->GetProgressGauge()) return;
-      dialog->GetProgressGauge()->SetValue(v);
-      dialog->GetProgressGauge()->Refresh();
-      dialog->GetProgressGauge()->Update();
-  };
-  dlg_ctx.set_range = [dialog](int r) {
-      if (!dialog || !dialog->GetProgressGauge()) return;
-      dialog->GetProgressGauge()->SetRange(r); };
-  dlg_ctx.pulse = [dialog](void) {
-      if (!dialog || !dialog->GetProgressGauge()) return;
-      dialog->GetProgressGauge()->Pulse(); };
-  dlg_ctx.set_message =
-      [dialog](const std::string& s) { dialog->SetMessage(wxString(s)); };
-  return dlg_ctx;
 }
 
 int RouteGui::SendToGPS(const wxString& com_name, bool bsend_waypoints,
